requirements.txt
pymupdf
pandas
openpyxl
tabula-py
jpype1